create the new data model
