+++ smartscope/objects/__init__.py
@@ -26,4 +26,4 @@
 # *
 # **************************************************************************
 
-from .data import *
+from .data_deprecated import *
